backend/requirements.txt
fastapi>=0.95.0
pydantic>=1.10.0
sqlalchemy>=1.4.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-dotenv>=0.19.0
alembic>=1.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
pytest>=7.0.0
httpx>=0.24.0
requests>=2.31.0
strawberry-graphql>=0.200.0 